Evaluate performance of DC methods on simulations

Quantify the performance of a differential co-expression pipeline on simulated data.

dcEvaluate(simulation, dclist, truth.type = c("association", "influence",
  "direct"), perf.method = "f.measure", combine = TRUE, ...)

Arguments

simulation
a list, storing data and results generated from simulations

dclist
a list of igraphs, produced using dcPipeline

truth.type
a character, specifying which level of the true network to retrieve: 'association' (default), 'influence' or 'direct'

perf.method
a character, specifying the method to use. Available methods can be accessed using perfMethods

combine
a logical, indicating whether differential networks from independent knock-outs should be treated as a single inference or independent inferences (defaults to TRUE)

...
additional parameters to be passed on to the performance metric method (see performanceMeasure)

Value

a numeric, representing the performance metric. A single value if combine = TRUE and a named vector otherwise.

See also

Examples

data(sim102)

#run a standard pipeline
resStd <- dcPipeline(sim102, dc.func = 'zscore')
dcEvaluate(sim102, resStd)

#> [1] 0.5625966

dcEvaluate(sim102, resStd, combine = FALSE)

#>      ADR1      UME6 
#> 0.7228916 0.3936508