dcanr: Differential co-expression/association network analysis

Application and evaluation of methods to infer differential co-expression/associations. An evaluation framework for testing differential co-expression methods in R. Simulated expression data with differential associations is used to assess inference methods. The package provides methods to evaluate inference methods and visualise predictions. Inference methods without implementations have been added to the package along with interfaces to existing implementations to enable ease of application and evaluations.

Details

There are three categories of functions available

Differential co-expression methods (DC) - These functions are used to perform a differential co-expression analysis on experimental data with binary conditions.

Functions to evaluate DC methods - These functions are used to evaluate methods implemented in the package and novel methods on simulated data. Expression data is simulated for 2 conditions, wild-type and knock-down of given genes.

By-products of implementations

Differential co-expression methods (DC)

dcMethods

dcScore

dcTest

dcAdjust

dcNetwork

Functions to evaluate DC methods

Accessors of simulated data:

getConditionNames

getSimData

getTrueNetwork

plotSimNetwork

Functions for evaluating inference methods

dcPipeline

dcEvaluate

By-products of implementations

These are functions used in the package but have further uses in general:

cor.pairs - a faster implementation of pairwise correlation computation

mi.ap - pairwise computation of mutual information MI with data discretisation performed using adaptive partitioning

perfMethods - available performance metrics

performanceMeasure - performance measures of prediction algorithms. Predictions have to be binary

See also